restructure router module
